Split up resource.py so we can use different parts as necessary. Update payment API structure a little bit. Move payment stuff from NameRequest resources into Payment resources. Add NRO update to complete payment procedure.
