+++ api/namex/resources/configuration.py
