+++ api/namex/resources/name_requests/__init__.py
@@ -1,2 +1,2 @@
-from .name_request import NameRequest
-from .name_requests import NameRequests
+from .name_request import NameRequestResource
+from .name_requests import NameRequestsResource
