+++ api/namex/resources/payment/payment.py
@@ -1,7 +1,7 @@
 import json
 
-from flask import request, make_response, jsonify
-from flask_restplus import Resource, cors, fields, marshal_with, marshal
+from flask import current_app, request, make_response, jsonify
+from flask_restplus import Resource, cors, fields
 from flask_jwt_oidc import AuthError
 
 from namex.utils.logging import setup_logging
@@ -10,10 +10,12 @@ from namex.utils.api_resource import clean_url_path_param, handle_exception
 
 from namex.constants import PaymentState, PaymentStatusCode
 
-from namex.models import Request as RequestDAO, Payment as PaymentDAO
+from namex.models import Request as RequestDAO, Payment as PaymentDAO, State
 
+from namex.resources.name_requests.abstract_nr_resource import AbstractNameRequestResource
 from namex.services.payment.exceptions import SBCPaymentException, SBCPaymentError, PaymentServiceError
-from namex.services.payment.payments import get_payment, create_payment, update_payment, CreatePaymentRequest, UpdatePaymentRequest
+from namex.services.payment.payments import get_payment, create_payment, update_payment
+from namex.services.name_request.utils import has_active_payment, get_active_payment
 
 from .api_namespace import api as payment_api
 from .utils import build_payment_request, merge_payment_request
@@ -120,105 +122,11 @@ def handle_auth_error(ex):
     return response
 
 
-@cors_preflight('GET, PUT')
-@payment_api.route('/<string:payment_identifier>', strict_slashes=False, methods=['GET', 'PUT', 'OPTIONS'])
+@cors_preflight('GET, POST')
+@payment_api.route('/<int:nr_id>', strict_slashes=False, methods=['GET', 'POST', 'OPTIONS'])
 @payment_api.doc(params={
-    'payment_identifier': ''
 })
-class Payment(Resource):
-    @staticmethod
-    @cors.crossdomain(origin='*')
-    # @jwt.requires_auth
-    @payment_api.response(200, 'Success', '')
-    # @marshal_with(payment_response_schema)
-    def get(payment_identifier):
-        try:
-            payment_identifier = clean_url_path_param(payment_identifier)
-
-            payment = get_payment(payment_identifier)
-
-            if not payment:
-                return jsonify(message=MSG_NOT_FOUND), 404
-
-            data = jsonify(payment.to_dict())
-            response = make_response(data, 200)
-            return response
-
-        except PaymentServiceError as err:
-            return handle_exception(err, err.message, 500)
-        except SBCPaymentException as err:
-            return handle_exception(err, err.message, err.status_code)
-        except SBCPaymentError as err:
-            return handle_exception(err, err.message, 500)
-        except Exception as err:
-            return handle_exception(err, err, 500)
-
-    @staticmethod
-    @cors.crossdomain(origin='*')
-    # @jwt.requires_auth
-    @payment_api.expect(payment_request_schema)
-    @payment_api.response(200, 'Success', '')
-    # @marshal_with()
-    def put(payment_identifier):
-        try:
-            payment_identifier = clean_url_path_param(payment_identifier)
-
-            json_input = request.get_json()
-            if not json_input:
-                return jsonify(message=MSG_BAD_REQUEST_NO_JSON_BODY), 400
-
-            # Grab the info we need off the request
-            payment_info = json_input.get('paymentInfo')
-            filing_info = json_input.get('filingInfo')
-            business_info = json_input.get('businessInfo')
-
-            # Update our payment request
-            req = PaymentRequest(
-                payment_info=payment_info,
-                filing_info=filing_info,
-                business_info=business_info
-            )
-
-            payment_response = update_payment(payment_identifier, req)
-            if not payment_response:
-                raise PaymentServiceError(message=MSG_ERROR_CREATING_RESOURCE)
-
-            data = jsonify(payment_response.to_dict())
-            response = make_response(data, 200)
-            return response
-
-        except PaymentServiceError as err:
-            return handle_exception(err, err.message, 500)
-        except SBCPaymentException as err:
-            return handle_exception(err, err.message, err.status_code)
-        except SBCPaymentError as err:
-            return handle_exception(err, err.message, 500)
-        except Exception as err:
-            return handle_exception(err, err, 500)
-
-
-@cors_preflight('DELETE')
-@payment_api.route('/extra', strict_slashes=False, methods=['DELETE', 'OPTIONS'])
-class Extra(Resource):
-    @staticmethod
-    @cors.crossdomain(origin='*')
-    # @jwt.requires_auth
-    # @payment_api.expect()
-    @payment_api.response(200, 'Success', '')
-    # @marshal_with()
-    @payment_api.doc(params={
-        '': ''
-    })
-    def delete():
-        pass
-
-
-@cors_preflight('GET')
-@payment_api.route('/<int:nr_id>/payments', strict_slashes=False, methods=['GET', 'OPTIONS'])
-@payment_api.doc(params={
-    'nr_id': 'NR Number - This field is required'
-})
-class NameRequestPayment(Resource):
+class NameRequestPayments(AbstractNameRequestResource):
     @cors.crossdomain(origin='*')
     def get(self, nr_id):
         try:
@@ -255,13 +163,6 @@ class NameRequestPayment(Resource):
         except Exception as err:
             return handle_exception(err, err, 500)
 
-
-@cors_preflight('GET, POST')
-@payment_api.route('/<int:nr_id>', strict_slashes=False, methods=['POST', 'OPTIONS'])
-@payment_api.doc(params={
-})
-class NameRequestPayments(Resource):
-    @staticmethod
     @cors.crossdomain(origin='*')
     # @jwt.requires_auth
     @payment_api.expect(payment_request_schema)
@@ -270,7 +171,16 @@ class NameRequestPayments(Resource):
     @payment_api.doc(params={
         'nr_id': 'Name Request number'
     })
-    def post(nr_id):
+    def post(self, nr_id):
+        """
+        At this point, the Name Request will still be using a TEMPORARY NR number.
+        Confirming the payment on the frontend triggers this endpoint. Here, we:
+        - Save the request to NRO which gives us a real NR.
+        - Create the payment via SBC Pay.
+        - If payment creation is successful, create a corresponding payment record in our system.
+        :param nr_id:
+        :return:
+        """
         try:
             # Find the existing name request
             nr_model = RequestDAO.query.get(nr_id)
@@ -278,6 +188,10 @@ class NameRequestPayments(Resource):
             if not nr_model:
                 # Should this be a 400 or 404... hmmm
                 return None, None, jsonify(message='{nr_id} not found'.format(nr_id=nr_id)), 400
+
+            # Save back to NRO to get the updated NR Number
+            update_solr = False
+            nr_model = self.update_records_in_network_services(nr_model, update_solr)
 
             json_input = request.get_json()
             payment_request = {}
@@ -346,13 +260,12 @@ class NameRequestPayments(Resource):
     'nr_id': '',
     'payment_id': ''
 })
-class NameRequestPayment(Resource):
-    @staticmethod
+class NameRequestPayment(AbstractNameRequestResource):
     @cors.crossdomain(origin='*')
     # @jwt.requires_auth
     @payment_api.response(200, 'Success', '')
     # @marshal_with(payment_response_schema)
-    def get(nr_id, payment_id):
+    def get(self, nr_id, payment_id):
         try:
             # Find the existing name request
             nr_model = RequestDAO.query.get(nr_id)
@@ -392,13 +305,12 @@ class NameRequestPayment(Resource):
         except Exception as err:
             return handle_exception(err, err, 500)
 
-    @staticmethod
     @cors.crossdomain(origin='*')
     # @jwt.requires_auth
     @payment_api.expect(payment_request_schema)
     @payment_api.response(200, 'Success', '')
     # @marshal_with()
-    def put(nr_id, payment_id):
+    def put(self, nr_id, payment_id):
         try:
             # Find the existing name request
             nr_model = RequestDAO.query.get(nr_id)
@@ -441,3 +353,93 @@ class NameRequestPayment(Resource):
             return handle_exception(err, err.message, 500)
         except Exception as err:
             return handle_exception(err, err, 500)
+
+
+@cors_preflight('PATCH')
+@payment_api.route('/<int:nr_id>/complete/<int:payment_id>', strict_slashes=False, methods=['PATCH', 'OPTIONS'])
+@payment_api.doc(params={
+    'nr_id': 'NR Number - This field is required'
+})
+class NameRequestCompletePayment(AbstractNameRequestResource):
+    # REST Method Handlers
+    @cors.crossdomain(origin='*')
+    def patch(self, nr_id, payment_id):
+        """
+        :param nr_id:
+        :param payment_id:
+        :return:
+        """
+        try:
+            # Find the existing name request
+            nr_model = RequestDAO.query.get(nr_id)
+            nr_svc = self.nr_service
+
+            nr_svc.nr_num = nr_model.nrNum
+            nr_svc.nr_id = nr_model.id
+
+            valid_update_states = [State.DRAFT, State.COND_RESERVE, State.RESERVED]
+
+            # This could be moved out, but it's fine here for now
+            def validate_patch_request(nr):
+                is_valid = True
+                msg = ''
+                if nr.stateCd in valid_update_states:
+                    is_valid = True
+
+                return is_valid, msg
+
+            is_valid_patch, validation_msg = validate_patch_request(nr_model)
+            validation_msg = validation_msg if not len(validation_msg) > 0 else 'Invalid request for PATCH'
+
+            if not is_valid_patch:
+                raise PaymentServiceError(message=validation_msg)
+
+            process_payment = has_active_payment(nr_model, payment_id)
+            if nr_model.stateCd in valid_update_states and not process_payment:
+                pass
+            elif process_payment:
+                # This handles updates if the NR state is DRAFT, COND_RESERVE or RESERVED
+                # If the state is COND_RESERVE update state to CONDITIONAL
+                # If the state is RESERVED update state to APPROVED
+                # Then update the name request as required
+                nr_model = self.process_payment(nr_model, payment_id)
+
+            current_app.logger.debug(nr_model.json())
+            response_data = nr_model.json()
+            # Add the list of valid Name Request actions for the given state to the response
+            response_data['actions'] = nr_svc.current_state_actions
+            return jsonify(response_data), 200
+        except PaymentServiceError as err:
+            return handle_exception(err, err.message, 500)
+        except Exception as err:
+            return handle_exception(err, repr(err), 500)
+
+    def process_payment(self, nr_model, payment_id):
+        nr_svc = self.nr_service
+
+        # Update the state of the payment
+        payment = get_active_payment(nr_model, payment_id)
+        sbc_payment_response = get_payment(payment.payment_token)
+
+        # TODO: Throw errors if this fails!
+        if sbc_payment_response.status_code == PaymentStatusCode.COMPLETED.value:
+            payment.payment_status_code = PaymentState.COMPLETED.value
+            payment.payment_completion_date = sbc_payment_response.created_on
+            payment.save_to_db()
+
+            # Use apply_state_change to change state, as it enforces the State change pattern
+            # apply_state_change takes the model, updates it to the specified state, and executes the callback handler
+            if nr_model.stateCd == State.DRAFT:
+                # If the state is DRAFT, leave it as a DRAFT
+                nr_model = nr_svc.apply_state_change(nr_model, State.DRAFT, self.handle_nr_approval)
+            if nr_model.stateCd == State.COND_RESERVE:
+                # If the state is COND_RESERVE update state to CONDITIONAL, and update the name request as required
+                nr_model = nr_svc.apply_state_change(nr_model, State.CONDITIONAL, self.handle_nr_approval)
+            elif nr_model.stateCd == State.RESERVED:
+                # If the state is RESERVED update state to APPROVED, and update the name request as required
+                nr_model = nr_svc.apply_state_change(nr_model, State.APPROVED, self.handle_nr_approval)
+
+            # This handles the updates for NRO and Solr, if necessary
+            self.update_records_in_network_services(nr_model)
+
+        return nr_model
