+++ api/namex/resources/payment/sbc_payment.py
@@ -1,0 +1,206 @@
+from flask import request, make_response, jsonify
+from flask_restplus import Resource, cors, fields
+from flask_jwt_oidc import AuthError
+
+from namex.utils.logging import setup_logging
+from namex.utils.auth import cors_preflight
+from namex.utils.api_resource import clean_url_path_param, handle_exception
+
+from namex.services.payment.exceptions import SBCPaymentException, SBCPaymentError, PaymentServiceError
+from namex.services.payment.payments import get_payment, update_payment
+
+from .api_namespace import api as payment_api
+
+from openapi_client.models import PaymentRequest
+
+setup_logging()  # It's important to do this first
+
+MSG_BAD_REQUEST_NO_JSON_BODY = 'No JSON data provided'
+MSG_SERVER_ERROR = 'Server Error!'
+MSG_NOT_FOUND = 'Resource not found'
+MSG_ERROR_CREATING_RESOURCE = 'Could not create / update resource'
+
+
+def validate_request(request):
+    return True
+
+
+# Define our DTO models
+# Generic model types
+dictionary_list_model = payment_api.model('DictionaryList', {
+    'key': fields.String,
+    'list': fields.List(fields.String)
+})
+
+dict_list_model = payment_api.model('DictionaryListList', {
+    'data': fields.List(fields.Nested(dictionary_list_model))
+})
+
+list_model = payment_api.model('List', {
+    'data': fields.List(fields.String)
+})
+
+string_model = payment_api.model('String', {
+    'data': fields.String
+})
+
+# Custom model types
+payment_info_schema = payment_api.model('PaymentInfo', {
+    'methodOfPayment': fields.String
+})
+
+filing_type_schema = payment_api.model('FilingType', {
+    'filingTypeCode': fields.String,
+    'priority': fields.Boolean,
+    'filingDescription': fields.String
+})
+
+filing_info_schema = payment_api.model('FilingInfo', {
+    'corpType': fields.String,
+    'date': fields.String,
+    'filingTypes': fields.List(fields.Nested(filing_type_schema)),
+})
+
+contact_info_schema = payment_api.model('ContactInfo', {
+    # 'firstName': fields.String,
+    # 'lastName': fields.String,
+    'addressLine1': fields.String,
+    'city': fields.String,
+    'province': fields.String,
+    'country': fields.String,
+    'postalCode': fields.String,
+})
+
+business_info_schema = payment_api.model('BusinessInfo', {
+    'businessIdentifier': fields.String,
+    'businessName': fields.String,
+    'contactInfo': fields.Nested(contact_info_schema)
+})
+
+
+payment_invoice_schema = payment_api.model('PaymentInvoice', {
+    'id': fields.String,
+    'referenceNumber': fields.String,
+    'statusCode': fields.String,
+    'createdBy': fields.String,
+    'createdOn': fields.String
+})
+
+payment_response_schema = payment_api.model('Payment', {
+    'id': fields.String,
+    'invoices': fields.List(fields.Nested(payment_invoice_schema)),
+    'paymentMethod': fields.String,
+    'statusCode': fields.String,
+    'createdBy': fields.String,
+    'createdOn': fields.String,
+    'updatedBy': fields.String,
+    'updatedOn': fields.String
+})
+
+# Define our request objects
+# These are POSTED use camelCase
+payment_request_schema = payment_api.model('PaymentRequest', {
+    'paymentInfo': fields.Nested(payment_info_schema),
+    'businessInfo': fields.Nested(business_info_schema),
+    'filingInfo': fields.Nested(filing_info_schema)
+})
+
+
+@payment_api.errorhandler(AuthError)
+def handle_auth_error(ex):
+    response = jsonify(ex.error)
+    response.status_code = ex.status_code
+    return response
+
+
+@cors_preflight('GET, PUT')
+@payment_api.route('/sbc-pay/<string:payment_identifier>', strict_slashes=False, methods=['GET', 'PUT', 'OPTIONS'])
+@payment_api.doc(params={
+    'payment_identifier': ''
+})
+class SBCPayment(Resource):
+    @staticmethod
+    @cors.crossdomain(origin='*')
+    # @jwt.requires_auth
+    @payment_api.response(200, 'Success', '')
+    # @marshal_with(payment_response_schema)
+    def get(payment_identifier):
+        try:
+            payment_identifier = clean_url_path_param(payment_identifier)
+
+            payment = get_payment(payment_identifier)
+
+            if not payment:
+                return jsonify(message=MSG_NOT_FOUND), 404
+
+            data = jsonify(payment.to_dict())
+            response = make_response(data, 200)
+            return response
+
+        except PaymentServiceError as err:
+            return handle_exception(err, err.message, 500)
+        except SBCPaymentException as err:
+            return handle_exception(err, err.message, err.status_code)
+        except SBCPaymentError as err:
+            return handle_exception(err, err.message, 500)
+        except Exception as err:
+            return handle_exception(err, err, 500)
+
+    @staticmethod
+    @cors.crossdomain(origin='*')
+    # @jwt.requires_auth
+    @payment_api.expect(payment_request_schema)
+    @payment_api.response(200, 'Success', '')
+    # @marshal_with()
+    def put(payment_identifier):
+        try:
+            payment_identifier = clean_url_path_param(payment_identifier)
+
+            json_input = request.get_json()
+            if not json_input:
+                return jsonify(message=MSG_BAD_REQUEST_NO_JSON_BODY), 400
+
+            # Grab the info we need off the request
+            payment_info = json_input.get('paymentInfo')
+            filing_info = json_input.get('filingInfo')
+            business_info = json_input.get('businessInfo')
+
+            # Update our payment request
+            req = PaymentRequest(
+                payment_info=payment_info,
+                filing_info=filing_info,
+                business_info=business_info
+            )
+
+            payment_response = update_payment(payment_identifier, req)
+            if not payment_response:
+                raise PaymentServiceError(message=MSG_ERROR_CREATING_RESOURCE)
+
+            data = jsonify(payment_response.to_dict())
+            response = make_response(data, 200)
+            return response
+
+        except PaymentServiceError as err:
+            return handle_exception(err, err.message, 500)
+        except SBCPaymentException as err:
+            return handle_exception(err, err.message, err.status_code)
+        except SBCPaymentError as err:
+            return handle_exception(err, err.message, 500)
+        except Exception as err:
+            return handle_exception(err, err, 500)
+
+
+@cors_preflight('DELETE')
+@payment_api.route('/sbc-pay/extra', strict_slashes=False, methods=['DELETE', 'OPTIONS'])
+class SBCPaymentExtra(Resource):
+    @staticmethod
+    @cors.crossdomain(origin='*')
+    # @jwt.requires_auth
+    # @payment_api.expect()
+    @payment_api.response(200, 'Success', '')
+    # @marshal_with()
+    @payment_api.doc(params={
+        '': ''
+    })
+    def delete():
+        pass
